structuring controllers
